pi/requirements.txt
requests>=2.32.0
python-dotenv>=1.0.0
gpiozero>=2.0
RPi.GPIO
pigpio
